refactor: upd userservice & wordservice
